refactor: include notification container into the overlay stack (#8198)
